feat(drivers): persist driver bins and repos across container recreation
Currently when core container is recreated, all repos need to be cloned and drivers need to be recompiled again from scratch. This can take >10mins and during this time the core instance is down. Many end users have reported this behaviour as a bug as it is unexpected to them. What would be the downside to persisting repos and driver bins across container recreation?